+++ dos/pdcscrn.c
@@ -669,7 +669,7 @@ int PDC_color_content(int color, int *red, int *green, int *blue)
     return OK;
 }
 
-int PDC_init_color(short color, short red, short green, short blue)
+int PDC_init_color(int color, int red, int green, int blue)
 {
     PDCREGS regs;
