+++ os2/pdcscrn.c
@@ -385,7 +385,7 @@ int PDC_color_content(int color, int *red, int *green, int *blue)
 #endif
 }
 
-int PDC_init_color(short color, short red, short green, short blue)
+int PDC_init_color(int color, int red, int green, int blue)
 {
 #ifdef PDCTHUNK
     THUNKEDVIO vcr;
