+++ x11/pdcscrn.c
@@ -717,7 +717,7 @@ int PDC_color_content(int color, int *red, int *green, int *blue)
     return OK;
 }
 
-int PDC_init_color(short color, short red, short green, short blue)
+int PDC_init_color(int color, int red, int green, int blue)
 {
     XColor tmp;
